+++ R/excel_dates.R
@@ -8,23 +8,57 @@
 #'
 #' @param date_num numeric vector of serial numbers to convert.
 #' @param date_system the date system, either \code{"modern"} or \code{"mac pre-2011"}.
-#' @return Returns a vector of class Date.
+#' @param include_time Include the time (hours, minutes, seconds) in the output? (See details)
+#' @param round_seconds Round the seconds to an integer (only has an effect when \code{include_time} is \code{TRUE})?
+#' @return Returns a vector of class Date if \code{include_time} is \code{FALSE}.  Returns a vector of class POSIXlt if \code{include_time} is \code{TRUE}.
+#' @details When using \code{include_time=TRUE} the returned object will not
+#'   include a time zone since excel does not store dates and times with time
+#'   zones.  When adding time zones, ensure that any conversion done does not
+#'   change the intended time including if the time is or is not in daylight
+#'   savings time.  Also, days with leap seconds will not be accurately handled
+#'   as they do not appear to be accurately handled by Windows (as described in
+#'   https://support.microsoft.com/en-us/help/2722715/support-for-the-leap-second).
 #' @export
 #' @examples
 #' excel_numeric_to_date(40000)
+#' excel_numeric_to_date(40000.5) # No time is included
+#' excel_numeric_to_date(40000.5, include_time=TRUE) # Time is included
+#' excel_numeric_to_date(40000.521, include_time=TRUE) # Time is included
+#' excel_numeric_to_date(40000.521, include_time=TRUE, round_seconds=FALSE) # Time with fractional seconds is included
 
 # Converts a numeric value like 42414 into a date "2016-02-14"
 
-excel_numeric_to_date <- function(date_num, date_system = "modern") {
+excel_numeric_to_date <- function(date_num, date_system = "modern", include_time=FALSE, round_seconds=TRUE) {
   if (!is.numeric(date_num)) {
     stop("argument `date_num` must be of class numeric")
   }
 
-  if (date_system == "mac pre-2011") {
-    as.Date(date_num, origin = "1904-01-01")
-  } else if (date_system == "modern") {
-    as.Date(date_num, origin = "1899-12-30")
-  } else {
-    stop("argument 'created' must be one of 'mac pre-2011' or 'modern'")
+  ret <-
+    if (date_system == "mac pre-2011") {
+      as.Date(floor(date_num), origin = "1904-01-01")
+    } else if (date_system == "modern") {
+      as.Date(floor(date_num), origin = "1899-12-30")
+    } else {
+      stop("argument 'created' must be one of 'mac pre-2011' or 'modern'")
+    }
+  if (include_time) {
+    ret <- as.POSIXlt(ret)
+    total_sec <- 86400*(date_num %% 1)
+    ret$sec <- total_sec %% 60
+    if (round_seconds) {
+      ret$sec <- round(ret$sec)
+    }
+    ret$min <- floor(total_sec/60) %% 60
+    ret$hour <- floor(total_sec/3600)
+    # Excel does not store the timezone; remove all ways of storing timezone and
+    # clean related fields (isdst).  Multiple methods are used by R for storing
+    # timezone information in POSIXlt objects
+    # (https://bugs.r-project.org/bugzilla3/show_bug.cgi?id=17437), and an
+    # attempt is made to clear all of them.
+    attr(ret, "tzone") <- NULL
+    ret$isdst <- 0L
+    ret$zone <- NULL
+    ret$gmtoff <- NULL
   }
+  ret
 }