+++ NEWS.md
@@ -1,7 +1,8 @@
 # janitor 1.0.1.9000  (current development version)
 
-Current dev't version on GitHub.
+## Major Features
 
+The new function `row_to_names` allows promoting a row of data to names.  And, thanks to **@billdenney** for writing this feature.
 
 # janitor 1.0.0  (2018-03-17)