+++ NEWS.md
@@ -1,7 +1,8 @@
 # janitor 1.0.1.9000  (current development version)
 
-Current dev't version on GitHub.
+## Major Features
 
+The new function `row_to_names` handles the case where a dirty data file is read in with its names stored as a row of the data.frame, rather than in the names.  This function sets the names of the data.frame to this row and optionally cleans up the rows above and including where the names were stored.  Thanks to **@billdenney** for writing this feature.
 
 # janitor 1.0.0  (2018-03-17)